+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `USE_TAKES` (default: `True`) config to be able to turn off takes packages.
+
 ## [1.1] 2016-08-29
 
 ### Added